Smart zoom columns of text with gesture/shortcut
PDFs usually have large margins which waste screen space. This feature
allows users to quickly zoom a column of text to fill the width of the
window by pressing Z key while hovering / double tap on touch devices.

This will also snap scroll to the horizontal axis when zoomed, to keep
the column in view while scrolling on a touchscreen/trackpad.